move amp-date-display
