+++ build-system/compile/bundles.config.bento.json
@@ -9,6 +9,15 @@
     }
   },
   {
+    "name": "amp-date-display",
+    "version": "1.0",
+    "latestVersion": "0.1",
+    "options": {
+      "npm": true,
+      "bento": true
+    }
+  },
+  {
     "name": "amp-jwplayer",
     "version": "1.0"
   }
