+++ build-system/compile/bundles.config.extensions.json
@@ -313,7 +313,6 @@
     "version": "1.0",
     "latestVersion": "0.1",
     "options": {
-      "npm": true,
       "bento": true,
       "hasCss": true
     }
