+++ build-system/test-configs/forbidden-terms.js
@@ -729,8 +729,8 @@ const forbiddenTermsGlobal = {
       'build-system/tasks/storybook/amp-env/main.js',
       'build-system/tasks/storybook/preact-env/main.js',
       'extensions/amp-animation/0.1/storybook/template.js',
-      'extensions/amp-date-display/1.0/storybook/Basic.amp.js',
-      'extensions/amp-date-display/1.0/storybook/Basic.js',
+      'src/bento/components/amp-date-display/1.0/storybook/Basic.amp.js',
+      'src/bento/components/amp-date-display/1.0/storybook/Basic.js',
       'extensions/amp-iframe/0.1/storybook/Basic.amp.js',
       'extensions/amp-iframe/1.0/storybook/Basic.amp.js',
       'extensions/amp-image-slider/0.1/storybook/Basic.amp.js',
