+++ src/bento/components/amp-date-display/1.0/component.type.js
