+++ src/bento/components/amp-date-display/1.0/storybook/Basic.amp.js
