+++ src/bento/components/amp-date-display/1.0/test/test-component.js
