+++ src/bento/components/amp-date-display/format.js
